script: Move DOMJSClass and get_dom_class to script_bindings.
